+++ src/main/java/me/quadradev/api/ApiApplication.java
@@ -1,11 +1,18 @@
 package me.quadradev.api;
 
-import me.quadradev.api.common.DotenvInitializer;
+import me.quadradev.api.infrastructure.config.DotenvInitializer;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.Banner;
 
-@SpringBootApplication
+@SpringBootApplication(
+        scanBasePackages = {
+                "me.quadradev.api.domain",
+                "me.quadradev.api.application",
+                "me.quadradev.api.infrastructure",
+                "me.quadradev.api.adapters"
+        }
+)
 public class ApiApplication {
 
     public static void main(String[] args) {
