+++ src/main/java/me/quadradev/api/adapters/incoming/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Incoming adapters such as controllers or event handlers.
+ */
+package me.quadradev.api.adapters.incoming;
