+++ src/main/java/me/quadradev/api/adapters/outgoing/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Outgoing adapters for repositories or external clients.
+ */
+package me.quadradev.api.adapters.outgoing;
