+++ src/main/java/me/quadradev/api/adapters/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Adapters layer containing incoming and outgoing interfaces.
+ */
+package me.quadradev.api.adapters;
