+++ src/main/java/me/quadradev/api/application/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Application layer: use cases and services.
+ */
+package me.quadradev.api.application;
