+++ src/main/java/me/quadradev/api/domain/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Domain layer: contains entities and ports.
+ */
+package me.quadradev.api.domain;
