+++ src/main/java/me/quadradev/api/infrastructure/config/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Configuration utilities for the infrastructure layer.
+ */
+package me.quadradev.api.infrastructure.config;
