+++ src/main/java/me/quadradev/api/infrastructure/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Infrastructure layer: technical implementations.
+ */
+package me.quadradev.api.infrastructure;
